thử fp8 emb

+++ OhMai/embedding.py
@@ -122,7 +122,7 @@ class OhMaiEmbedding(nn.Module):
         self.active_tokens = None # Cần kích hoạt mỗi lần forward
 
         if active_vocab is None: active_vocab = vocab // 2  # a safe assumption
-        w = torch.empty(active_vocab, self.hidim, device="cuda", dtype=self.weight.dtype)
+        w = torch.empty(active_vocab, self.hidim, device="cuda", dtype=torch.float8_e4m3fn)
         self.active_weight = nn.Parameter(w)
         self.active_vocab = active_vocab
